made APAgent docker_run.sh use AP_IFACE environment variable

+++ APAgent/docker_run.sh
@@ -1,9 +1,17 @@
 #!/bin/sh
 
+source approject/.env.local
+AP_IFACE="$AP_IFACE"
+
+if [ -z "$AP_IFACE" ]; then
+    echo "(Error) AP_IFACE not set in 'approject/.env.local'."
+    exit 1
+fi
+
 sudo docker-compose down
 
 sudo docker system prune -f
 
-nmcli dev set wlan0 managed no
+nmcli dev set $AP_IFACE managed no
 
 sudo docker-compose up --build
